+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Finalised the basic features of the avatar editor.
+
 ## [0.12.0-0] - 2026-01-16
 
 ## Added